pyproject.toml
[project]
name = "agent-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=3.4.3",
    "dotenv>=0.9.9",
    "google-adk>=1.1.1",
    "google-genai>=1.18.0",
    "litellm>=1.71.3",
]

[dependency-groups]
dev = [
    "ruff>=0.11.12",
]
